+++ packages/widgetbook/lib/src/fields/num_slider_field.dart
@@ -7,7 +7,7 @@ class NumSliderField<T extends num> extends Field<T> {
     required super.name,
     this.divisions,
     super.initialValue,
-    @deprecated super.onChanged,
+    @Deprecated('Fields should not be aware of their context') super.onChanged,
     required this.min,
     required this.max,
     required super.codec,